Add the ability to set AWS.S3 options.
This PR adds the ability to set options on the root `AWS.S3` object.
I also did a little surgery on the `makefile` + `package.json` that
got the `make test` target to work correctly on my machine. LMK
if any of the tweaks aren't aligned with where you'd like to take
things!